requirements.txt
Django==1.3
argparse==1.2.1
demjson==1.6
django-dajax==0.9.2
django-dajaxice==0.5.5
django-haystack==2.0.0-beta
django-mongodb-engine==0.4.0
django-tastypie==0.9.11
djangotoolbox==0.9.2
gunicorn==0.17.4
haystack==0.10
mimeparse==0.1.3
numpy==1.7.0
odict==1.5.1
pymongo==2.4.2
pystache==0.3.1
python-dateutil==1.0
python-memcached==1.48
python-ptrace==0.6.4
simplejson==3.1.0
wsgiref==0.1.2
yolk==0.4.3
